api/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
pybase62==0.6.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
uuid0==0.2.7
